backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytesseract
transformers
torch
Pillow
